i'm dumb

+++ app/models/bid.rb
@@ -20,19 +20,23 @@ class Bid < ActiveRecord::Base
 
   def update_remaining_salary
     old_bid = self.player.top_bid(self.team.league_id)
-    old_team = old_bid.team
-    old_team.remaining_salary += old_bid.amount
-    old_team.save
-    old_bid.destroy
+    if old_bid
+      old_team = old_bid.team
+      old_team.remaining_salary += old_bid.amount
+      old_team.save
+      old_bid.destroy
+    end
 
     self.team.remaining_salary -= self.amount - (amount_was || 0)
     self.team.save
   end
 
   def can_place_bid
+    old_bid = self.player.top_bid(self.team.league_id)
+
     if self.team.remaining_salary < self.amount - (amount_was || 0)
       errors.add(:amount, 'cannot exceed remaining salary')
-    elsif self.amount <= self.player.top_bid(self.team.league_id).amount
+    elsif old_bid && self.amount <= old_bid.amount
       errors.add(:amount, 'must be greater than current bid')
     end
   end
